Switch AmendmentBlockHandler to Repeat
