+++ src/etl/impl/AmendmentBlockHandler.cpp
@@ -1,7 +1,7 @@
 //------------------------------------------------------------------------------
 /*
     This file is part of clio: https://github.com/XRPLF/clio
-    Copyright (c) 2023, the clio developers.
+    Copyright (c) 2024, the clio developers.
 
     Permission to use, copy, modify, and distribute this software for any
     purpose with or without fee is hereby granted, provided that the above
@@ -17,33 +17,40 @@
 */
 //==============================================================================
 
-#include "etl/SystemState.hpp"
 #include "etl/impl/AmendmentBlockHandler.hpp"
-#include "util/AsioContextTestFixture.hpp"
-#include "util/MockPrometheus.hpp"
+
+#include "etl/SystemState.hpp"
+#include "util/log/Logger.hpp"
 
 #include <boost/asio/io_context.hpp>
-#include <gmock/gmock.h>
-#include <gtest/gtest.h>
 
 #include <chrono>
-#include <cstddef>
+#include <functional>
+#include <utility>
 
-using namespace etl::impl;
+namespace etl::impl {
 
-struct AmendmentBlockHandlerTest : util::prometheus::WithPrometheus, SyncAsioContextTest {
-    testing::StrictMock<testing::MockFunction<void()>> actionMock;
-    etl::SystemState state;
+AmendmentBlockHandler::ActionType const AmendmentBlockHandler::defaultAmendmentBlockAction = []() {
+    static util::Logger const log{"ETL"};
+    LOG(log.fatal()) << "Can't process new ledgers: The current ETL source is not compatible with the version of "
+                     << "the libxrpl Clio is currently using. Please upgrade Clio to a newer version.";
 };
 
-TEST_F(AmendmentBlockHandlerTest, CallToOnAmendmentBlockSetsStateAndRepeatedlyCallsAction)
+AmendmentBlockHandler::AmendmentBlockHandler(
+    boost::asio::io_context& ioc,
+    SystemState& state,
+    std::chrono::steady_clock::duration interval,
+    ActionType action
+)
+    : state_{std::ref(state)}, repeat_{ioc}, interval_{interval}, action_{std::move(action)}
 {
-    AmendmentBlockHandler handler{ctx, state, std::chrono::nanoseconds{1}, actionMock.AsStdFunction()};
-
-    EXPECT_FALSE(state.isAmendmentBlocked);
-    EXPECT_CALL(actionMock, Call()).Times(testing::AtLeast(10));
-    handler.onAmendmentBlock();
-    EXPECT_TRUE(state.isAmendmentBlocked);
-
-    runContextFor(std::chrono::milliseconds{1});
 }
+
+void
+AmendmentBlockHandler::onAmendmentBlock()
+{
+    state_.get().isAmendmentBlocked = true;
+    repeat_.start(interval_, action_);
+}
+
+}  // namespace etl::impl
