+++ tests/common/util/AsioContextTestFixture.hpp
@@ -26,6 +26,7 @@
 #include <boost/asio/spawn.hpp>
 #include <gmock/gmock.h>
 
+#include <chrono>
 #include <optional>
 #include <thread>
 
@@ -95,6 +96,14 @@ struct SyncAsioContextTest : virtual public NoLoggerFixture {
     {
         ctx.run();
         ctx.reset();
+    }
+
+    void
+    runContextFor(std::chrono::milliseconds duration)
+    {
+        ctx.run_for(duration);
+        ctx.stop();
+        // ctx.reset() is not used because it will reset stopped flag
     }
 
 protected:
